--- TCC/index.html
@@ -1,58 +1,0 @@
-<html>
-
-<head>
-    <title>TCC</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0-beta/css/materialize.min.css">
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0-beta/js/materialize.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/vue/dist/vue.js"></script>
-
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta charset="UTF-8">
-
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <!--teste-->
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.2/js/materialize.min.js"></script>
-
-
-
-</head>
-
-<body>
-    <script type="text/javascript">
-        $('.modal').modal();
-    </script>
-
-
-
-
-    <div id="telaPrincipal">
-        <span v-show="seen" v-html="template"></span>
-    </div>
-    <div id="novoOuImportar">
-        <span v-show="seen" v-html="template1"></span>
-    </div>
-    <div id="nodeOuJs">
-        <span v-show="seen" v-html="template2"></span>
-    </div>
-
-    <script src="novoOuImportar.js"></script>
-    <script src="nodeOuJs.js"></script>
-    
-    <script src="telaPrincipal.js"></script>
-
-    <script src="https://unpkg.com/vue"></script>
-
-    <script type="text/javascript">
-        $('.modal').modal();
-    </script>
-    <script src="acoes.js"></script>
-
-
-</body>
-
-</html>
